+++ frontend/src/Components/Organisms/PlayBackScreen/PlaybackScreen.tsx
@@ -5,6 +5,7 @@ import { PauseCircleOutlineOutlined } from '@mui/icons-material';
 import * as THREE from 'three';
 import GeneModel from '../../../Utilities/GeneModel/GeneModel'
 import GeneModelStorage from '../../../Utilities/GeneModel/GeneModelStorage' 
+import MeshStorage from "../../../Utilities/Mesh/MeshStorage";
 import GeneEffectPlayer from '../../../Utilities/GeneEffects/GeneEffectPlayer'
 import Recorder from '../Recorder'
 import CanvasSize from '../../../Utilities/GlobalVarriables/CanvasSize'
@@ -12,6 +13,7 @@ import { Vector2 } from "three";
 
 interface Props {
   geneModelStorage    : GeneModelStorage;
+  meshStorage         : MeshStorage;
   canvasSize          : CanvasSize;
   isPlayingFlg        : boolean;
   reqInstPlayFlg      : boolean;
@@ -169,10 +171,13 @@ export class PlaybackScreen extends React.Component<Props, State>{
       * Imp: すべてのMeshをシーンに追加する
       */
       this.props.geneModelStorage.storage.forEach( (geneModel: any) => {
-        this.state.threeScene.add(geneModel.mesh);
+        const targetMesh = this.props.meshStorage.getMeshById(geneModel.id);
+        if (targetMesh){
+          this.state.threeScene.add(targetMesh.mesh);
+        }
       })
     }
-
+    //a
  
   playBackThree = () => {
      /**
@@ -186,7 +191,7 @@ export class PlaybackScreen extends React.Component<Props, State>{
            
            // Effectを発火
            this.props.geneModelStorage.storage.forEach( (geneModel: GeneModel) => {
-               GeneEffectPlayer.play(geneModel);
+               GeneEffectPlayer.play(this.props.meshStorage, geneModel);
            })
  
            this.state.threeRenderer.render(this.state.threeScene, this.state.threeCamera);
